Changing aggregate version to long and Segregating Order infrastructu…

+++ src/Services/Catalog/Application/Abstractions/EventSourcing/EventStore/Events/Snapshot.cs
@@ -6,7 +6,7 @@ public abstract record Snapshot<TAggregate, TId>
     where TAggregate : IAggregateRoot<TId>, new()
     where TId : struct
 {
-    public int AggregateVersion { get; init; }
+    public long AggregateVersion { get; init; }
     public TId AggregateId { get; init; }
     public string AggregateName { get; } = typeof(TAggregate).Name;
     public TAggregate AggregateState { get; init; } = new();
